\name{is.number}
\alias{is.number}

\title{ Test for a number }

\description{ Internal CoCoCore function: 
              Returns TRUE if its argument is of type real or
              type integer and not is a missing value indicator,
              and FALSE otherwise.
}

\usage{
is.number(x)
}

\arguments{
  \item{x}{ object to be tested. }
}


\details{ This function is not to be called by the user. }

\value{ Logical }

\references{ Badsberg, J.H.: Xlisp+CoCo, Aalborg University, 1996. }

\author{ Jens Henrik Badsberg }

% \note{ }

% \section{ToDo}{....}

\seealso{ \code{\link{is.numeric}}, \code{\link{is.complex}}, and \code{\link{is.na}}. }

% \examples{ }

\keyword{ interface }% at least one, from doc/KEYWORDS
\keyword{ multivariate }% __ONLY ONE__ keyword per line
