\name{my.not.a.number}
\alias{my.not.a.number}

\title{ The invalid number for CoCo objects }

\description{
  Internal CoCoCore function:
  The double send at init in \code{\link{coco.init}} to CoCo objects
  to be returned as the invalid number.
}

\usage{
my.not.a.number()
}

\details{ This function is not to be called by the user. }

% \value{
%   ~Describe the value returned
%   If it is a LIST, use
%   \item{comp1 }{Description of 'comp1'}
%   \item{comp2 }{Description of 'comp2'}
%   ...
% }

\references{ Badsberg, J.H.: Xlisp+CoCo, Aalborg University, 1996. }

\author{ Jens Henrik Badsberg }

% \note{ }

% \section{ToDo}{....}

\seealso{ \code{\link{coco.init}}. }

% \examples{ }

\keyword{ interface }% at least one, from doc/KEYWORDS
\keyword{ multivariate }% __ONLY ONE__ keyword per line
